deleting task which is assigned to a plan is no longer possible. AlertDialog is displayed. Test added.

+++ Friendly-plans/app/src/main/java/pg/autyzm/friendly_plans/manager_app/view/task_list/TaskListActivity.java
@@ -1,10 +1,16 @@
 package pg.autyzm.friendly_plans.manager_app.view.task_list;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+
+import java.util.List;
+
+import database.entities.PlanTaskTemplate;
 import database.repository.TaskTemplateRepository;
 import javax.inject.Inject;
 import pg.autyzm.friendly_plans.ActivityProperties;
@@ -27,11 +33,18 @@ public class TaskListActivity extends AppCompatActivity {
 
                 @Override
                 public void onRemoveTaskClick(int position){
-                    taskTemplateRepository.delete(taskListAdapter.getTaskItem(position).getId());
-                    toastUserNotifier.displayNotifications(
-                            R.string.task_removed_message,
-                            getApplicationContext());
-                    taskListAdapter.setTaskItems(taskTemplateRepository.getAll());
+                    List<PlanTaskTemplate> planTasks = taskTemplateRepository.getPlansWithThisTask(
+                            taskListAdapter.getTaskItem(position).getId());
+                    if(planTasks.isEmpty()) {
+                        taskTemplateRepository.delete(taskListAdapter.getTaskItem(position).getId());
+                        toastUserNotifier.displayNotifications(
+                                R.string.task_removed_message,
+                                getApplicationContext());
+                        taskListAdapter.setTaskItems(taskTemplateRepository.getAll());
+                    }
+                    else{
+                        displayTaskCannotBeRemovedAlert(planTasks);
+                    }
                 }
 
                 @Override
@@ -67,5 +80,29 @@ public class TaskListActivity extends AppCompatActivity {
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         taskListAdapter = new TaskRecyclerViewAdapter(taskItemClickListener);
         recyclerView.setAdapter(taskListAdapter);
+    }
+
+    private void displayTaskCannotBeRemovedAlert(List<PlanTaskTemplate> planTasks) {
+        String planNames = "";
+        for (PlanTaskTemplate planTask : planTasks){
+            String asd = planTask.getPlanTemplate().getName();
+            planNames = planNames + asd + ", ";
+        }
+        planNames = planNames.substring(0, planNames.length() - 2);
+        planNames += ".";
+        AlertDialog alertDialog = new AlertDialog.Builder(
+                TaskListActivity.this).create();
+        alertDialog.setTitle(R.string.break_time_tasks);
+        alertDialog.setMessage(
+                getResources().getString(R.string.task_cannot_be_removed_message)
+                        + " " + planNames);
+        alertDialog.setButton(AlertDialog.BUTTON_POSITIVE,
+                getResources().getString(R.string.task_cannot_be_removed_dialog_close_button_text),
+                new DialogInterface.OnClickListener() {
+                    public void onClick(DialogInterface dialog, int which) {
+                        dialog.dismiss();
+                    }
+                });
+        alertDialog.show();
     }
 }
